- Switched to absolute imports. - Restructured to Daemon/Service specific Utilities. Common utils are held under src/Common. - Implemented catching of PyMineException in Communicator.
Aside from addServer and removeServer (for which I do not currently have a strategy) this commit of the daemon half of the coin is incredibly stable. I could not come up with a way to cause poor behavior when interacting with start/stop/restart/use/arbitrary. However, a full test suite should be written up.